requirements.txt
python-telegram-bot>=20.3
telegram>=0.0.1
mlx>=0.20.0
mlx-lm>=0.20.0
httpx>=0.27.0
websockets>=14.0
web3>=7.0.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
alembic>=1.13.0
pandas>=2.2.0
numpy>=2.0.0
scipy>=1.14.0
rich>=13.9.0
textual>=0.85.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.9.0
python-dateutil>=2.9.0
pytz>=2024.2
aiofiles>=24.1.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
psutil>=6.1.0
